+++ src/com/company/Sphere.java
@@ -3,23 +3,23 @@ package com.company;
 public class Sphere extends Hitable {
 
     Vec3 center;
-    float radius;
+    double radius;
     Material material;
 
-    Sphere(Vec3 cen, float r, Material m) {
+    Sphere(Vec3 cen, double r, Material m) {
         this.center = cen;
         this.radius = r;
         this.material = m;
     }
 
-    boolean hit(Ray r, float t_min, float t_max, HitRecord rec) {
+    boolean hit(Ray r, double t_min, double t_max, HitRecord rec) {
         Vec3 oc = Vec3.vec_minus(r.origin(), center);
-        float a = Vec3.dot(r.direction(), r.direction());
-        float b = Vec3.dot(oc, r.direction());
-        float c = Vec3.dot(oc, oc) - radius * radius;
-        float discriminant = b * b - a * c;
+        double a = Vec3.dot(r.direction(), r.direction());
+        double b = Vec3.dot(oc, r.direction());
+        double c = Vec3.dot(oc, oc) - radius * radius;
+        double discriminant = b * b - a * c;
         if (discriminant > 0) {
-            float temp = (float) ((-b - Math.sqrt(discriminant))/a);
+            double temp = ((-b - Math.sqrt(discriminant))/a);
             if (temp < t_max && temp > t_min) {
                 rec.t = temp;
                 rec.p = r.point_at_parameter(rec.t);
@@ -28,7 +28,7 @@ public class Sphere extends Hitable {
                 rec.material = material;
                 return true;
             }
-            temp = (float) ((-b + Math.sqrt(discriminant))/a);
+            temp = ((-b + Math.sqrt(discriminant))/a);
             if (temp < t_max && temp > t_min) {
                 rec.t = temp;
                 rec.p = r.point_at_parameter(rec.t);
